Reduce memory consumption by storing FFT on dataGrid
