+++ library/Mag/DataIO.py
@@ -1,6 +1,6 @@
 import numpy as np
 import scipy as sp
-from . import Simulator
+from . import (Simulator, MathUtils)
 from scipy.spatial import cKDTree
 from SimPEG.Utils import mkvc
 import matplotlib.pyplot as plt
@@ -11,6 +11,225 @@ import os
 from shapely.geometry import mapping, LineString
 import fiona
 from fiona.crs import from_epsg
+
+
+class dataGrid(object):
+    """
+        Grid data object
+    """
+
+    x0, y0 = 0., 0.
+    nx, ny = 1, 1
+    dx, dy = 1., 1.
+    values = None
+    inc = 90.
+    dec = 90.
+
+    def __init__(self, **kwargs):
+        for key, value in kwargs.items():
+            setattr(self, key, value)
+
+        return
+
+    @property
+    def hx(self):
+
+        if getattr(self, '_hx', None) is None:
+            self._hx = np.asarray(range(self.nx)) * self.dx + self.x0
+
+        return self._hx
+
+    @property
+    def hy(self):
+
+        if getattr(self, '_hy', None) is None:
+            self._hy = np.asarray(range(self.ny)) * self.dy + self.y0
+
+        return self._hy
+
+    @property
+    def npadx(self):
+
+        if getattr(self, '_npadx', None) is None:
+            self._npadx = int(np.floor(self.values.shape[1]))
+
+        return self._npadx
+
+    @property
+    def npady(self):
+
+        if getattr(self, '_npady', None) is None:
+            self._npady = int(np.floor(self.values.shape[0]))
+
+        return self._npady
+
+    @property
+    def gridPadded(self):
+
+        if getattr(self, '_gridPadded', None) is None:
+            # Add paddings
+            dpad = np.c_[
+                np.fliplr(self.values[:, 0:self.npadx]),
+                self.values,
+                np.fliplr(self.values[:, -self.npadx:])
+                ]
+
+            dpad = np.r_[
+                np.flipud(dpad[0:self.npady, :]),
+                dpad,
+                np.flipud(dpad[-self.npady:, :])
+                ]
+
+            # Tapper the paddings
+            rampx = -np.cos(np.pi*np.asarray(range(self.npadx))/self.npadx)
+            rampx = np.r_[rampx, np.ones(self.values.shape[1]), -rampx]/2. + 0.5
+            # tapperx,_ = meshgrid(rampx,np.ones(dpad.shape[1]))
+            # tapperx[padx:-padx,:] = 1.
+
+            rampy = -np.cos(np.pi*np.asarray(range(self.npady))/self.npady)
+            rampy = np.r_[rampy, np.ones(self.values.shape[0]), -rampy]/2. + 0.5
+            tapperx, tappery = np.meshgrid(rampx, rampy)
+
+            self._gridPadded = tapperx*tappery*dpad
+
+        return self._gridPadded
+
+    @property
+    def Kx(self):
+
+        if getattr(self, '_Kx', None) is None:
+
+            dx = (self.dx)/(self.gridPadded.shape[1]-1)
+            dy = (self.dy)/(self.gridPadded.shape[0]-1)
+
+            kx = np.fft.fftfreq(self.gridPadded.shape[1], dx)
+            ky = np.fft.fftfreq(self.gridPadded.shape[0], dy)
+
+            Ky, Kx = np.meshgrid(ky, kx)
+            self._Ky, self._Kx = Ky.T, Kx.T
+
+        return self._Kx
+
+    @property
+    def Ky(self):
+
+        if getattr(self, '_Ky', None) is None:
+
+            dx = (self.dx)/(self.gridPadded.shape[1]-1)
+            dy = (self.dy)/(self.gridPadded.shape[0]-1)
+
+            kx = np.fft.fftfreq(self.gridPadded.shape[1], dx)
+            ky = np.fft.fftfreq(self.gridPadded.shape[0], dy)
+
+            Ky, Kx = np.meshgrid(ky, kx)
+            self._Ky, self._Kx = Ky.T, Kx.T
+
+        return self._Ky
+
+    @property
+    def gridFFT(self):
+
+        if getattr(self, '_gridFFT', None) is None:
+
+            self._gridFFT = np.fft.fft2(self.gridPadded)
+
+        return self._gridFFT
+
+    @property
+    def derivativeX(self):
+
+        if getattr(self, '_derivativeX', None) is None:
+
+            FHxD = (self.Kx*1j)*self.gridFFT
+            fhxd_pad = np.fft.ifft2(FHxD)
+            self._derivativeX = np.real(
+                fhxd_pad[self.npady:-self.npady, self.npadx:-self.npadx])
+
+        return self._derivativeX
+
+    @property
+    def derivativeY(self):
+
+        if getattr(self, '_derivativeY', None) is None:
+
+            FHyD = (self.Ky*1j)*self.gridFFT
+
+            fhyd_pad = np.fft.ifft2(FHyD)
+            self._derivativeY = np.real(
+                fhyd_pad[self.npady:-self.npady, self.npadx:-self.npadx])
+
+        return self._derivativeY
+
+    @property
+    def firstVertical(self):
+
+        if getattr(self, '_firstVertical', None) is None:
+
+            FHzD = self.gridFFT*np.sqrt(self.Kx**2. + self.Ky**2.)
+            fhzd_pad = np.fft.ifft2(FHzD)
+            self._firstVertical = np.real(
+                fhzd_pad[self.npady:-self.npady, self.npadx:-self.npadx])
+
+        return self._firstVertical
+
+    @property
+    def totalHorizontal(self):
+
+        if getattr(self, '_totalHorizontal', None) is None:
+
+            self._totalHorizontal = np.sqrt(self.derivativeX**2. + self.derivativeY**2. + 1e-8)
+
+        return self._totalHorizontal
+
+    @property
+    def tiltAngle(self):
+
+        if getattr(self, '_tiltAngle', None) is None:
+
+            Ftilt = self.gridFFT*np.sqrt(self.Kx**2. + self.Ky**2. + 1e-8)
+            tilt_pad = np.fft.ifft2(Ftilt)
+            tilt = np.real(
+                tilt_pad[self.npady:-self.npady, self.npadx:-self.npadx])
+            self._tiltAngle = np.arctan2(tilt, self.totalHorizontal)
+
+        return self._tiltAngle
+
+    @property
+    def analyticSignal(self):
+
+        if getattr(self, '_analyticSignal', None) is None:
+
+            self._analyticSignal = np.sqrt(self.derivativeX**2. + self.derivativeY**2. + self.firstVertical**2. + 1e-8)
+
+        return self._analyticSignal
+
+    @property
+    def RTP(self):
+
+        if getattr(self, '_RTP', None) is None:
+
+            h0_xyz = MathUtils.dipazm_2_xyz(self.inc, self.dec)
+            Frtp = self.gridFFT / (
+                (h0_xyz[2] + 1j*(self.Kx*h0_xyz[0] + self.Ky*h0_xyz[1]))**2.
+            )
+            rtp_pad = np.fft.ifft2(Frtp)
+            self._RTP = np.real(
+                rtp_pad[self.npady:-self.npady, self.npadx:-self.npadx])
+
+        return self._RTP
+
+    def upwardContinuation(self, z=0):
+        """
+            Function to calculate upward continued data
+        """
+
+        upFact = -np.sqrt(self.Kx**2. + self.Ky**2.)*z/np.sqrt(self.dx**2. + self.dy**2.)
+        FzUpw = self.gridFFT*np.exp(upFact)
+        zUpw_pad = np.fft.ifft2(FzUpw)
+        zUpw = np.real(
+            zUpw_pad[self.npady:-self.npady, self.npadx:-self.npadx])
+
+        return zUpw
 
 
 def loadGRDFile(fileName, plotIt=True):
@@ -88,38 +307,6 @@ def loadGeoTiffFile(fileName, plotIt=True):
 
     return data
 
-
-class dataGrid(object):
-    """
-        Grid data object
-    """
-
-    x0, y0 = 0., 0.
-    nx, ny = 1, 1
-    dx, dy = 1., 1.
-    values = None
-
-    def __init__(self, **kwargs):
-        for key, value in kwargs.items():
-            setattr(self, key, value)
-
-        return
-
-    @property
-    def hx(self):
-
-        if getattr(self, '_hx', None) is None:
-            self._hx = np.asarray(range(self.nx)) * self.dx + self.x0
-
-        return self._hx
-
-    @property
-    def hy(self):
-
-        if getattr(self, '_hy', None) is None:
-            self._hy = np.asarray(range(self.ny)) * self.dy + self.y0
-
-        return self._hy
 
 
 def arrayToRaster(array, fileName, EPSGCode, xMin, xMax, yMin, yMax, numBands, dataType='image'):
